requirements.txt
# Purpose: Minimal requirements for GPU tests using PyTorch.
# Latest Changes: Updated to use flexible versioning without pinning specific versions.
# Key Logic: Only install packages necessary to verify GPU functionality.
# Expected File Path: /project_root/requirements.txt
# Reasoning: Using flexible versioning to avoid conflicts with system-installed packages.

# Note: To install the CUDA-enabled builds, use the command:
#       pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu118

# Core dependencies with flexible versioning
torch
torchvision
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.5.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Additional dependencies from your working project
qdrant-client
openai-clip
streamlit
rawpy
exifread
transformers
pillow
accelerate
bitsandbytes

# Core video processing
opencv-python>=4.5.0
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# Transcription
openai-whisper>=20230314
# whisperx is optional and can be installed separately if needed

# Computer vision and embeddings
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
ftfy>=6.1.1
regex>=2022.1.18
transformers>=4.23.0
git+https://github.com/openai/CLIP.git

# Vector database
# qdrant-client>=1.1.1 (optional, can be installed separately)

# Utilities
tqdm>=4.65.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0

# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
torch>=2.0.0  # CUDA support
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
ffmpeg-python>=0.2.0

# Audio processing
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.0

# Vision models
open_clip_torch>=2.20.0
salesforce-lavis>=1.0.0  # For BLIP

# Video processing
scenedetect>=0.6.0
av>=10.0.0  # PyAV for efficient video handling

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Optional: for development
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
